Moved list views into list.py
