+++ corehq/apps/export/urls.py
@@ -7,29 +7,31 @@ from corehq.apps.export.views.download import (
     DownloadNewCaseExportView,
     DownloadNewSmsExportView,
 )
+from corehq.apps.export.views.list import (
+    DailySavedExportListView,
+    FormExportListView,
+    CaseExportListView,
+    DashboardFeedListView,
+    DeIdFormExportListView,
+    DeIdDailySavedExportListView,
+    DeIdDashboardFeedListView,
+)
 from corehq.apps.export.old_views import (
     CreateNewCustomFormExportView,
     CreateNewCustomCaseExportView,
     EditNewCustomFormExportView,
     EditNewCustomCaseExportView,
     DeleteNewCustomExportView,
-    FormExportListView,
-    CaseExportListView,
-    DeIdFormExportListView,
     GenerateSchemaFromAllBuildsView,
     download_daily_saved_export,
-    DashboardFeedListView,
     CreateNewCaseFeedView,
     CreateNewFormFeedView,
     EditCaseFeedView,
     EditCaseDailySavedExportView,
     EditFormFeedView,
     EditFormDailySavedExportView,
-    DailySavedExportListView,
     CreateNewDailySavedCaseExport,
     CreateNewDailySavedFormExport,
-    DeIdDailySavedExportListView,
-    DeIdDashboardFeedListView,
     DashboardFeedPaywall,
     DailySavedExportPaywall,
     CopyExportView,
